AERO: To test that the `GET_SERVICES` query is functioning, I created a `RequestAppointment.jsx` page that queries the database to retrieve the services currently within the database. I also created a `AppointmentCheckboxInput.jsx` component that creates a checkbox input type for each entry.
I have added barebone components of a form that we will likely want to use for users to submit appointment requests. DOES NOT currently submit appointments. Must troubleshoot managing state to accurately reflect which services a user has selected.

The form has the following endpoint `/request-appointment` and is incorporated in the React Router in `main.jsx` but must be added to the nav bar.